Add PDB import function to file manager namespace
